Rectify check 11

+++ 0x15-file_io/0-read_textfile.c
@@ -42,7 +42,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	}
 	str[bytes_read] = '\0';
 	bytes_written = write(0, str, bytes_read);
-	if (bytes_written == -1 || (size_t)bytes_written != (size_t)bytes_read)
+	if (bytes_written == -1 || bytes_written != bytes_read)
 	{
 	close(fd);
 	free(str);
